Move AbstractConnection#readMessage() and #writeMessage() into SocketConnection and WebRTCConnection as each is only used in one place
