+++ clientserver/src/main/java/net/rptools/clientserver/simple/connection/SocketConnection.java
@@ -17,6 +17,7 @@ package net.rptools.clientserver.simple.connection;
 import java.io.*;
 import java.net.Socket;
 import java.net.SocketTimeoutException;
+import net.rptools.clientserver.ActivityListener;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -118,7 +119,7 @@ public class SocketConnection extends AbstractConnection implements Connection {
           }
 
           try {
-            SocketConnection.this.writeMessage(out, message);
+            writeMessage(out, message);
           } catch (IOException e) {
             log.error("Error while writing message. Closing connection.", e);
             return;
@@ -127,6 +128,29 @@ public class SocketConnection extends AbstractConnection implements Connection {
       } finally {
         SocketConnection.this.close();
       }
+    }
+
+    protected final void writeMessage(OutputStream out, byte[] message) throws IOException {
+      int length = message.length;
+
+      notifyListeners(ActivityListener.Direction.Outbound, ActivityListener.State.Start, length, 0);
+
+      out.write(length >> 24);
+      out.write(length >> 16);
+      out.write(length >> 8);
+      out.write(length);
+
+      for (int i = 0; i < message.length; i++) {
+        out.write(message[i]);
+
+        if (i != 0 && i % ActivityListener.CHUNK_SIZE == 0) {
+          notifyListeners(
+              ActivityListener.Direction.Outbound, ActivityListener.State.Progress, length, i);
+        }
+      }
+      out.flush();
+      notifyListeners(
+          ActivityListener.Direction.Outbound, ActivityListener.State.Complete, length, length);
     }
   }
 
@@ -154,7 +178,7 @@ public class SocketConnection extends AbstractConnection implements Connection {
 
         while (!SocketConnection.this.isClosed() && SocketConnection.this.isAlive()) {
           try {
-            byte[] message = SocketConnection.this.readMessage(in);
+            byte[] message = readMessage(in);
             SocketConnection.this.dispatchCompressedMessage(message);
           } catch (SocketTimeoutException e) {
             log.warn("Lost client {}", SocketConnection.this.getId(), e);
@@ -171,6 +195,33 @@ public class SocketConnection extends AbstractConnection implements Connection {
         SocketConnection.this.close();
         fireDisconnect();
       }
+    }
+
+    private byte[] readMessage(InputStream in) throws IOException {
+      int b32 = in.read();
+      int b24 = in.read();
+      int b16 = in.read();
+      int b8 = in.read();
+
+      if (b32 < 0) {
+        throw new IOException("Stream closed");
+      }
+      int length = (b32 << 24) + (b24 << 16) + (b16 << 8) + b8;
+
+      notifyListeners(ActivityListener.Direction.Inbound, ActivityListener.State.Start, length, 0);
+
+      byte[] ret = new byte[length];
+      for (int i = 0; i < length; i++) {
+        ret[i] = (byte) in.read();
+
+        if (i != 0 && i % ActivityListener.CHUNK_SIZE == 0) {
+          notifyListeners(
+              ActivityListener.Direction.Inbound, ActivityListener.State.Progress, length, i);
+        }
+      }
+      notifyListeners(
+          ActivityListener.Direction.Inbound, ActivityListener.State.Complete, length, length);
+      return ret;
     }
   }
 }
